\name{getProxy-package}
\alias{getProxy-package}
\alias{getProxy}
\docType{package}
\title{
\packageTitle{getProxy}
}
\description{
\packageDescription{getProxy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{getProxy}
\packageIndices{getProxy}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{getProxy}

Maintainer: \packageMaintainer{getProxy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
